requirements/test.txt
-r base.txt
bleach==1.5.0
flake8==3.0.4
pycodestyle==2.1.0
pyflakes==1.3.0
pytest==3.0.3
pytest-cov==2.4.0
check-manifest==0.34
coverage==4.2
tox==2.4.1
coveralls==1.1
factory-boy==2.7.0
pytest-django==3.0.0
PyYAML==3.12
docopt==0.6.2
fake-factory==0.7.2
django-coverage-plugin==1.3.1
